Bump batch size from uint8_t to uint16_t

+++ src/send.h
@@ -19,11 +19,11 @@ typedef struct {
 	char* packets;
 	uint32_t* ips;
 	int* lens;
-	uint8_t len;
-	uint8_t capacity;
+	uint16_t len;
+	uint16_t capacity;
 }batch_t;
 
-batch_t* create_packet_batch(uint8_t capacity);
+batch_t* create_packet_batch(uint16_t capacity);
 void free_packet_batch(batch_t* batch);
 
 #endif // SEND_H
